test(math): move long/oversized equation coverage to the root matrix e2e and show the KaTeX example in all four demos
